une_support_system/ai_agents/serializers.py
from rest_framework import serializers
from .models import AiAgentsRequest


class AgentRequestSerializer(serializers.ModelSerializer):
    class Meta:
        model = AiAgentsRequest
        fields = ["message"]
